Déplacement index (+ reliquat)
